+++ simpleble/src/backends/macos/PeripheralBase.mm
@@ -1,5 +1,6 @@
 #import "PeripheralBase.h"
 #import "PeripheralBaseMacOS.h"
+#import "ServiceBuilder.h"
 
 #import "CommonUtils.h"
 
@@ -53,6 +54,14 @@ void PeripheralBase::update_advertising_data(advertising_data_t advertising_data
     is_connectable_ = advertising_data.connectable;
     manufacturer_data_ = advertising_data.manufacturer_data;
     rssi_ = advertising_data.rssi;
+
+    // Append services that haven't been seen before
+    for (auto& service : advertising_data.service_uuids) {
+        if (std::find(advertised_services_.begin(), advertised_services_.end(), service) ==
+            advertised_services_.end()) {
+            advertised_services_.push_back(service);
+        }
+    }
 }
 
 void PeripheralBase::connect() {
@@ -90,7 +99,12 @@ std::vector<Service> PeripheralBase::services() {
 }
 
 std::vector<Service> PeripheralBase::advertised_services() {
-    return {};
+    std::vector<Service> service_list;
+    for (auto& service_uuid : advertised_services_) {
+        service_list.push_back(ServiceBuilder(service_uuid));
+    }
+
+    return service_list;
 }
 
 std::map<uint16_t, ByteArray> PeripheralBase::manufacturer_data() { return manufacturer_data_; }